api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
# psycopg2-binary==2.9.9  # Commented out for local SQLite development
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
alembic==1.13.1
ruff==0.1.9
black==23.12.1
isort==5.13.2
openai==1.3.7
python-docx==1.1.0
python-docx-replace==0.1.0
Pillow==10.1.0
requests==2.31.0
